+++ src/json.hpp
@@ -70,6 +70,7 @@ SOFTWARE.
 #include "detail/parsing/binary_writer.hpp"
 #include "detail/serializer.hpp"
 #include "detail/json_ref.hpp"
+#include "adl_serializer.hpp"
 
 /*!
 @brief namespace for Niels Lohmann
@@ -78,42 +79,6 @@ SOFTWARE.
 */
 namespace nlohmann
 {
-template<typename, typename>
-struct adl_serializer
-{
-    /*!
-    @brief convert a JSON value to any value type
-
-    This function is usually called by the `get()` function of the
-    @ref basic_json class (either explicit or via conversion operators).
-
-    @param[in] j         JSON value to read from
-    @param[in,out] val  value to write to
-    */
-    template<typename BasicJsonType, typename ValueType>
-    static void from_json(BasicJsonType&& j, ValueType& val) noexcept(
-        noexcept(::nlohmann::from_json(std::forward<BasicJsonType>(j), val)))
-    {
-        ::nlohmann::from_json(std::forward<BasicJsonType>(j), val);
-    }
-
-    /*!
-    @brief convert any value type to a JSON value
-
-    This function is usually called by the constructors of the @ref basic_json
-    class.
-
-    @param[in,out] j  JSON value to write to
-    @param[in] val     value to read from
-    */
-    template<typename BasicJsonType, typename ValueType>
-    static void to_json(BasicJsonType& j, ValueType&& val) noexcept(
-        noexcept(::nlohmann::to_json(j, std::forward<ValueType>(val))))
-    {
-        ::nlohmann::to_json(j, std::forward<ValueType>(val));
-    }
-};
-
 class json_pointer
 {
     /// allow basic_json to access private members
@@ -7968,7 +7933,6 @@ inline bool operator!=(json_pointer const& lhs, json_pointer const& rhs) noexcep
     return not (lhs == rhs);
 }
 } // namespace nlohmann
-
 
 ///////////////////////
 // nonmember support //
